backend/requirements.txt
# ============================================================
#  FILE    : backend/requirements.txt
#  FUNGSI  : Dependencies backend v2
#  INSTALL : pip install -r requirements.txt
# ============================================================

# MQTT client
paho-mqtt==1.6.1

# PostgreSQL adapter
psycopg2-binary>=2.9.9

# REST API framework
fastapi==0.115.12

# ASGI server
uvicorn[standard]==0.34.0

# Environment variables
python-dotenv==1.0.1

# Data validation
pydantic==2.11.3